backend/requirements.txt
alembic==1.5.5
astroid==2.2.5
Authlib==0.15.3
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
Click==7.0
colorama==0.4.4
cryptography==3.4.6
ecdsa==0.13.2
Flask==1.0.2
Flask-Cors==3.0.8
Flask-Migrate==2.7.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.0
future==0.17.1
gunicorn==20.0.4
idna==2.10
isort==4.3.18
itsdangerous==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.4.0
Mako==1.1.4
MarkupSafe==1.1.1
mccabe==0.6.1
psycopg2-binary==2.8.6
pyasn1==0.4.8
pycparser==2.20
pylint==2.3.1
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
python-jose==3.2.0
requests==2.25.1
rsa==4.7.2
six==1.12.0
SQLAlchemy==1.3.3
typed-ast==1.4.1
urllib3==1.26.3
Werkzeug==1.0.1
wrapt==1.11.1
